requirements.txt
python-dotenv==1.0.0
pytz==2023.3
chromadb==0.4.22
groq==0.4.2
requests==2.31.0
rich==13.7.0
